aws-s3-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
boto3==1.36.7
pydantic-settings==2.7.1
python-dotenv==1.0.1
